+++ pymdownx/superfences.py
@@ -33,10 +33,8 @@ from markdown.preprocessors import Preprocessor
 from markdown.blockprocessors import CodeBlockProcessor
 from markdown.extensions.attr_list import get_attrs
 from markdown import util as md_util
-from .util import warn_deprecated
 import functools
 import re
-from inspect import signature
 
 SOH = '\u0001'  # start
 EOT = '\u0004'  # end
@@ -169,15 +167,16 @@ def highlight_validator(language, inputs, options, attrs, md):
 
     for k, v in inputs.items():
         matched = False
-        for opt, validator in (('hl_lines', RE_HL_LINES), ('linenums', RE_LINENUMS)):
-            if k == opt and use_pygments:
-                matched = True
-                if v is True or validator.match(v) is None:
-                    attrs[k] = v
-                elif use_pygments:
-                    options[k] = v
-                break
-
+        if use_pygments:
+            if k.startswith('data-'):
+                attrs[k] = v
+                continue
+            for opt, validator in (('hl_lines', RE_HL_LINES), ('linenums', RE_LINENUMS), ('title', None)):
+                if k == opt:
+                    if v is not True and (validator is None or validator.match(v) is not None):
+                        options[k] = v
+                        matched = True
+                        break
         if not matched:
             attrs[k] = v
 
@@ -192,17 +191,11 @@ def default_validator(language, inputs, options, attrs, md):
     return True
 
 
-def _validator(language, inputs, options, attrs, md, validator=None, _legacy=False):
+def _validator(language, inputs, options, attrs, md, validator=None):
     """Validator wrapper."""
 
     md.preprocessors['fenced_code_block'].get_hl_settings()
-    if _legacy:
-        value = validator(language, inputs)
-        for k, v in inputs.items():
-            options[k] = v
-        return value
-    else:
-        return validator(language, inputs, options, attrs, md)
+    return validator(language, inputs, options, attrs, md)
 
 
 def _formatter(src='', language='', options=None, md=None, class_name="", _fmt=None, **kwargs):
@@ -227,7 +220,6 @@ class SuperFencesCodeExtension(Extension):
         self.config = {
             'disable_indented_code_blocks': [False, "Disable indented code blocks - Default: False"],
             'custom_fences': [[], 'Specify custom fences. Default: See documentation.'],
-            'highlight_code': [True, "Deprecated and does nothing"],
             'css_class': [
                 '',
                 "Set class name for wrapper element. The default of CodeHilite or Highlight will be used"
@@ -278,19 +270,11 @@ class SuperFencesCodeExtension(Extension):
             class_name = custom.get('class')
             fence_format = custom.get('format', fence_code_format)
             validator = custom.get('validator', default_validator)
-            legacy = False
             if name is not None and class_name is not None:
-                sig = signature(validator)
-                if len(sig.parameters) == 2:
-                    legacy = True
-                    warn_deprecated(
-                        "Old format of custom validators is deprectated, please migrate to the new format"
-                        ": validator(language, inputs, options, attrs, md)"
-                    )
                 self.extend_super_fences(
                     name,
                     functools.partial(_formatter, class_name=class_name, _fmt=fence_format),
-                    functools.partial(_validator, validator=validator, _legacy=legacy)
+                    functools.partial(_validator, validator=validator)
                 )
 
         self.md = md
@@ -370,14 +354,11 @@ class SuperFencesBlockPreprocessor(Preprocessor):
 
         if not self.checked_hl_settings:
             self.checked_hl_settings = True
-            if not self.config['highlight_code']:
-                warn_deprecated(
-                    "Disabling of 'highlight_code' is deprecated and no longer does anything.",
-                )
 
             config = None
             self.highlighter = None
             for ext in self.md.registeredExtensions:
+                self.highlight_ext = ext
                 try:
                     config = getattr(ext, "get_pymdownx_highlight_settings")()
                     self.highlighter = getattr(ext, "get_pymdownx_highlighter")()
@@ -399,9 +380,13 @@ class SuperFencesBlockPreprocessor(Preprocessor):
             self.linenums_style = config.get('linenums_style', 'table')
             self.linenums_class = config.get('linenums_class', 'linenums')
             self.linenums_special = config.get('linenums_special', -1)
-            self.wrapcode = not config.get('legacy_no_wrap_code', False)
             self.language_prefix = config.get('language_prefix', 'language-')
             self.code_attr_on_pre = config.get('code_attr_on_pre', False)
+            self.auto_title = config.get('auto_title', False)
+            self.auto_title_map = config.get('auto_title_map', {})
+            self.line_spans = config.get('line_spans', '')
+            self.line_anchors = config.get('line_anchors', '')
+            self.anchor_linenums = config.get('anchor_linenums', False)
 
     def clear(self):
         """Reset the class variables."""
@@ -753,6 +738,7 @@ class SuperFencesBlockPreprocessor(Preprocessor):
         linestart = None
         linespecial = None
         hl_lines = None
+        title = None
 
         if self.use_pygments:
             if 'hl_lines' in options:
@@ -765,11 +751,16 @@ class SuperFencesBlockPreprocessor(Preprocessor):
                 linestep = m.group('linestep')
                 linespecial = m.group('linespecial')
                 del options['linenums']
+            if 'title' in options:
+                title = options['title']
+                del options['title']
 
         linestep = self.parse_line_step(linestep)
         linestart = self.parse_line_start(linestart)
         linespecial = self.parse_line_special(linespecial)
         hl_lines = self.parse_hl_lines(hl_lines)
+
+        self.highlight_ext.pygments_code_block += 1
 
         el = self.highlighter(
             guess_lang=self.guess_lang,
@@ -781,9 +772,13 @@ class SuperFencesBlockPreprocessor(Preprocessor):
             linenums_special=self.linenums_special,
             linenums_class=self.linenums_class,
             extend_pygments_lang=self.extend_pygments_lang,
-            wrapcode=self.wrapcode,
             language_prefix=self.language_prefix,
-            code_attr_on_pre=self.code_attr_on_pre
+            code_attr_on_pre=self.code_attr_on_pre,
+            auto_title=self.auto_title,
+            auto_title_map=self.auto_title_map,
+            line_spans=self.line_spans,
+            line_anchors=self.line_anchors,
+            anchor_linenums=self.anchor_linenums
         ).highlight(
             src,
             language,
@@ -794,7 +789,9 @@ class SuperFencesBlockPreprocessor(Preprocessor):
             linespecial=linespecial,
             classes=classes,
             id_value=id_value,
-            attrs=attrs
+            attrs=attrs,
+            title=title,
+            code_block_count=self.highlight_ext.pygments_code_block
         )
 
         return el